new file:   controllers/magnetController.js 	new file:   models/magnetModel.js 	new file:   routes/index.js 	new file:   views/infohash.pug 	deleted:    views/single.pug 	modified:   webserver.js
