--- views/single.pug
@@ -1,28 +1,0 @@
-html
-    include includes/header.pug
-    body
-        include includes/navbar.pug
-        .container
-            each item in result
-                .card
-                    .card-header
-                        h4=item.name
-                    .card-body
-                        p.card-text= 'Infohash: ' + item.infohash
-                        p.card-text= 'Fetched At: ' + new Date(item.fetchedAt)
-                        // - var filestring = item.files;
-                        // - var formatstring = filestring.split(',').join('\n');
-                        // pre= formatstring || item.name
-                        a.btn.btn-primary(href=item.magnet + trackers) Magnet URI
-            br
-            br
-
-            if health
-                p="Seeds: " + health.seeds
-                p="Peers: " + health.peers
-            br
-
-            h5= "Query Time: " + timer + "ms"
-            br
-            br
-            include includes/footer.pug
